Remove tdnn architecture from fluent speech commands recipe
